Deletion functionality for all admin pages (#6)
## Delete functionality
- change all DB relations to cascade on delete
- add delete functionality for `users & dozenten`, `kurse `and `studiengänge`
- added deletion routes
- added importable status message component
- added status messages to all admin pages
- added a check to stop last admin from being deleted

## Misc
- update composer & npm dependencies
- rename all kurse, users & studiengänge index files to `list_...`
- fixed wrong indices of `dozenten.sql`
- added more test users to test data

+++ app/Http/Controllers/ProfileController.php
@@ -3,9 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\ProfileUpdateRequest;
+use App\Models\Dozent;
+use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\View\View;
 
@@ -40,21 +43,46 @@ class ProfileController extends Controller
     /**
      * Delete the user's account.
      */
-    public function destroy(Request $request): RedirectResponse
+    public function destroy(Request $request, int $id, bool $isUser = true): RedirectResponse
     {
-        $request->validateWithBag('userDeletion', [
-            'password' => ['required', 'current_password'],
-        ]);
+        DB::beginTransaction();
 
-        $user = $request->user();
+        try {
+            if ($isUser) {
+                $user = User::findOrFail($id);
 
-        Auth::logout();
+                // check if user is last admin
+                if ($user->admin && User::where('admin', true)->count() === 1) {
+                    throw new \Exception('Der letzte Administrator kann nicht gelöscht werden.');
+                }
 
-        $user->delete();
+                // handle current user deletion
+                if (Auth::id() === $user->id) {
+                    Auth::logout();
+                    $request->session()->invalidate();
+                    $request->session()->regenerateToken();
+                }
 
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
+                if ($user->dozent) {
+                    $message = 'Benutzer & Dozent erfolgreich gelöscht.';
+                }
+                else {
+                    $message = 'Benutzer erfolgreich gelöscht.';
+                }
 
-        return Redirect::to('/');
+                $user->delete(); // cascade delete associated tables (e.g. dozenten, ...)
+            }
+            else {
+                $dozent = Dozent::findOrFail($id);
+                $dozent->delete(); // cascade delete associated tables (e.g. kurse, ...)
+                $message = 'Dozent erfolgreich gelöscht.';
+            }
+
+            DB::commit();
+            return redirect()->route('users.index')->with('success', $message);
+        } catch (\Exception $e) {
+            DB::rollBack();
+            return redirect()->back()->with('error', 'Es ist ein Fehler aufgetreten: ' . $e->getMessage());
+        }
     }
 }
